requirements.txt
psutil
pystray
matplotlib
pillow
